requirements.txt
langchain>=0.1.0
python-dotenv>=1.0.0
intercom-python>=4.0.0
chromadb>=0.4.0
tiktoken>=0.5.0
unstructured>=0.10.0
python-magic-bin>=0.4.14
pdf2image>=1.16.3
pytesseract>=0.3.10
opencv-python>=4.8.0
pillow>=10.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
